Migrate away from Stackblitz
Must now run project locally using npm
