--- index.html
@@ -1,1 +1,0 @@
-<div id="root"></div>
