+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "react-ts",
-  "version": "0.0.0",
+  "name": "dota_drafter",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
     "@babel/core": "^7.0.0",
@@ -8,7 +8,7 @@
     "@emotion/styled": "^11.3.0",
     "@mantine/core": "5.5.5",
     "@mantine/hooks": "5.5.5",
-    "@tabler/icons": "^1.110.0",
+    "@tabler/icons": "^1.111.0",
     "@types/react": "18.0.21",
     "@types/react-dom": "18.0.6",
     "react": "18.2.0",
@@ -25,5 +25,21 @@
   },
   "devDependencies": {
     "react-scripts": "latest"
+  },
+  "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
+  "main": "index.js",
+  "author": "",
+  "license": "ISC",
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
