+++ public/index.html
@@ -1,0 +1,1 @@
+<div id="root"></div>
