--- style.css
@@ -1,3 +1,0 @@
-/* html {
-  background-color: #f4ffff;
-} */
